refactor: review 관련 파일 옮기고, 필요없는 .gitkeep 파일 제거 (#114)
