+++ src/main/java/com/knoc/senior/SeniorProfileService.java
@@ -13,8 +13,7 @@ import com.knoc.senior.entity.SeniorProfile;
 import com.knoc.senior.entity.SeniorSkill;
 import com.knoc.senior.repository.SeniorProfileQueryRepository;
 import com.knoc.senior.repository.SeniorProfileRepository;
-import com.knoc.settlement.repository.ReviewFeedbackRepository;
-import lombok.Builder;
+import com.knoc.reviewFeedback.repository.ReviewFeedbackRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
