--- src/main/java/com/knoc/settlement/service/ReviewFeedbackService.java
@@ -1,65 +1,0 @@
-package com.knoc.settlement.service;
-
-import com.knoc.global.exception.BusinessException;
-import com.knoc.global.exception.ErrorCode;
-import com.knoc.member.Member;
-import com.knoc.member.MemberRepository;
-import com.knoc.order.entity.Order;
-import com.knoc.order.entity.OrderStatus;
-import com.knoc.order.repository.OrderRepository;
-import com.knoc.senior.entity.SeniorProfile;
-import com.knoc.senior.repository.SeniorProfileRepository;
-import com.knoc.settlement.dto.ReviewFeedbackRequestDto;
-import com.knoc.settlement.entity.ReviewFeedback;
-import com.knoc.settlement.repository.ReviewFeedbackRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class ReviewFeedbackService {
-
-    private final ReviewFeedbackRepository reviewFeedbackRepository;
-    private final OrderRepository orderRepository;
-    private final SeniorProfileRepository seniorProfileRepository;
-    private final MemberRepository memberRepository;
-
-    @Transactional
-    public void createReview(ReviewFeedbackRequestDto dto, Long juniorId) {
-        //1. 주문 확인
-        Order order = orderRepository.findById(dto.getOrderId()).orElseThrow(() -> new BusinessException(ErrorCode.ORDER_NOT_FOUND));
-        // 2. 요청자가 해당 주문의 주니어인지 검증
-        if(!order.getJunior().getId().equals(juniorId)){
-            throw new BusinessException(ErrorCode.NOT_JUNIOR_FOR_ORDER);
-        }
-        //3. 결제 완료 상태인지 확인 (결제 완료PAID 상태에서만 후기 작성이 가능)
-        if (order.getStatus() != OrderStatus.PAID) {
-            throw new BusinessException(ErrorCode.REVIEW_NOT_ALLOWED);
-        }
-        //4. 후기를 이미 작성한 주문인지 검증
-        if (reviewFeedbackRepository.existsByOrderId(order.getId())) {
-            throw new BusinessException(ErrorCode.REVIEW_ALREADY_EXISTS);
-        }
-        
-        //5. 시니어 프로필 확인
-        SeniorProfile senior = seniorProfileRepository.findByMemberId(order.getSenior().getId())
-                .orElseThrow(() -> new BusinessException(ErrorCode.SENIOR_PROFILE_NOT_FOUND));
-
-        //6. 검증완료-> 저장
-        ReviewFeedback review=ReviewFeedback.builder()
-                .order(order)
-                .junior(order.getJunior())
-                .seniorProfile(senior)
-                .rating(dto.getRating())
-                .comment(dto.getComment())
-                .build();
-        reviewFeedbackRepository.save(review);
-
-        //시니어 프로필 avgReview 최신화
-        senior.updateRating(dto.getRating());
-
-    }
-
-}
