+++ src/test/java/com/knoc/dashboard/DashboardServiceTest.java
@@ -11,8 +11,8 @@ import com.knoc.order.repository.OrderRepository;
 import com.knoc.senior.entity.SeniorProfile;
 import com.knoc.senior.entity.SeniorSkill;
 import com.knoc.senior.repository.SeniorProfileRepository;
-import com.knoc.settlement.entity.ReviewFeedback;
-import com.knoc.settlement.repository.ReviewFeedbackRepository;
+import com.knoc.reviewFeedback.entity.ReviewFeedback;
+import com.knoc.reviewFeedback.repository.ReviewFeedbackRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
