+++ src/test/java/com/knoc/settlement/service/ReviewFeedbackServiceTest.java
@@ -8,11 +8,12 @@ import com.knoc.member.MemberRepository;
 import com.knoc.order.entity.Order;
 import com.knoc.order.entity.OrderStatus;
 import com.knoc.order.repository.OrderRepository;
+import com.knoc.reviewFeedback.service.ReviewFeedbackService;
 import com.knoc.senior.entity.SeniorProfile;
 import com.knoc.senior.repository.SeniorProfileRepository;
-import com.knoc.settlement.dto.ReviewFeedbackRequestDto;
-import com.knoc.settlement.entity.ReviewFeedback;
-import com.knoc.settlement.repository.ReviewFeedbackRepository;
+import com.knoc.reviewFeedback.dto.ReviewFeedbackRequestDto;
+import com.knoc.reviewFeedback.entity.ReviewFeedback;
+import com.knoc.reviewFeedback.repository.ReviewFeedbackRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
